chore: consolidate protocol module
This patch consolidates the code in the `crate::protocol` module.
