+++ src/handlers/document_close.rs
@@ -1,6 +1,6 @@
 use crate::cache::Cache;
 use crate::handlers::{Error, RequestHandler};
-use crate::protocol::requests::{PolymorphicRequest, Request};
+use crate::protocol::{PolymorphicRequest, Request};
 
 use async_trait::async_trait;
