+++ src/handlers/document_symbol.rs
@@ -1,7 +1,6 @@
 use crate::cache::Cache;
 use crate::handlers::{Error, RequestHandler};
-use crate::protocol::requests::{PolymorphicRequest, Request};
-use crate::protocol::responses::Response;
+use crate::protocol::{PolymorphicRequest, Request, Response};
 use crate::shared::structs::RequestContext;
 use crate::visitors::semantic::utils;
 use crate::visitors::semantic::SymbolsVisitor;
