+++ src/handlers/folding.rs
@@ -1,7 +1,6 @@
 use crate::cache::Cache;
 use crate::handlers::{Error, RequestHandler};
-use crate::protocol::requests::{PolymorphicRequest, Request};
-use crate::protocol::responses::Response;
+use crate::protocol::{PolymorphicRequest, Request, Response};
 use crate::visitors::semantic::{utils, FoldFinderVisitor};
 
 use flux::semantic::walk::{self, Node};
