--- src/protocol/mod.rs
@@ -1,3 +1,0 @@
-pub mod notifications;
-pub mod requests;
-pub mod responses;
