--- src/protocol/responses.rs
@@ -1,34 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize, Clone)]
-pub struct Response<T> {
-    pub id: u32,
-    pub result: Option<T>,
-    pub jsonrpc: String,
-}
-
-impl<T> Response<T>
-where
-    T: Serialize + Clone,
-{
-    pub fn new(id: u32, result: Option<T>) -> Response<T> {
-        Response {
-            id,
-            result,
-            jsonrpc: String::from("2.0"),
-        }
-    }
-
-    pub fn to_json(&self) -> Result<String, String> {
-        match serde_json::to_string(self) {
-            Ok(s) => Ok(s),
-            Err(_) => {
-                Err("Failed to serialize initialize response"
-                    .to_string())
-            }
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone)]
-pub struct ShutdownResult {}
